feat: Implemented upsert and delete all in ParentDao and ParentRepository
The `Upsert` annotation in `ParentDao` was replaced with `Insert(onConflict = OnConflictStrategy.IGNORE)` to reflect its functionality as an upsert operation.

A `deleteAll` function was added to `ParentDao` to remove all entries from the `lists` table.

A corresponding `removeAll` function was added to `ParentRepository` to call the new `deleteAll` DAO function.