Update README with additional technologies and license restriction
Added Hilt, Room, Navigation Compose, and Material 3 to the list of technologies used in the project.
Updated the copyright year to 2025.
Clarified that item categories and labels are planned features.
Removed "Share list item" and "Copy list item" from key features.
Added a restriction to the license, prohibiting commercial use of the software.