Add share app functionality to Settings screen
Added a share icon button to the `ReusableTopAppBar` in `SettingsScreen.kt`.
When clicked, this button triggers an intent to share a link to the app on the Google Play Store.

The share icon button is implemented as a new `TopAppBarIcon` composable in `app/src/main/java/com/softklass/lazuli/ui/theme/Iconography.kt`.
This composable uses `FilledIconButton` with `MaterialTheme.colorScheme.primaryContainer` for the container color and `MaterialTheme.colorScheme.onPrimaryContainer` for the content color.
The share icon itself is `Icons.Rounded.Share`.